Lower lambdas and generators directly in ExprRewriter
