+++ src/transform/expr.rs
@@ -1,11 +1,12 @@
 use super::{
     context::Context, rewrite_assert, rewrite_class_def, rewrite_complex_expr, rewrite_decorator,
-    rewrite_expr_to_stmt, rewrite_for_loop, rewrite_import, rewrite_match_case, rewrite_string,
-    rewrite_try_except, rewrite_with, Options,
+    rewrite_for_loop, rewrite_import, rewrite_match_case, rewrite_string, rewrite_try_except,
+    rewrite_with, Options,
 };
 use crate::body_transform::{walk_expr, walk_stmt, Transformer};
 use crate::template::{make_binop, make_generator, make_tuple, make_unaryop, single_stmt};
 use crate::{py_expr, py_stmt};
+use ruff_python_ast::name::Name;
 use ruff_python_ast::{self as ast, Expr, Operator, Stmt, UnaryOp};
 use ruff_text_size::TextRange;
 use std::mem::take;
@@ -39,17 +40,109 @@ __dp__.truth({expr:expr})
         );
     }
 
-    fn lower_lambdas_generators(&mut self, stmt: &mut Stmt) -> bool {
-        let mut lowerer = rewrite_expr_to_stmt::LambdaGeneratorLowerer::new(self.ctx);
-        lowerer.rewrite(stmt);
-        let mut lowered_functions = lowerer.into_statements();
-        if lowered_functions.is_empty() {
-            false
+    fn lower_lambda(&mut self, lambda: ast::ExprLambda) -> Expr {
+        let func_name = self.ctx.fresh("lambda");
+
+        let ast::ExprLambda {
+            parameters, body, ..
+        } = lambda;
+
+        let parameters = parameters
+            .map(|params| *params)
+            .unwrap_or_else(|| ast::Parameters {
+                range: TextRange::default(),
+                node_index: ast::AtomicNodeIndex::default(),
+                posonlyargs: vec![],
+                args: vec![],
+                vararg: None,
+                kwonlyargs: vec![],
+                kwarg: None,
+            });
+
+        let func_def = py_stmt!(
+            "\ndef {func:id}():\n    return {body:expr}",
+            func = func_name.as_str(),
+            body = *body,
+        );
+
+        let func_def = match func_def {
+            Stmt::FunctionDef(mut function_def) => {
+                function_def.parameters = Box::new(parameters);
+                Stmt::FunctionDef(function_def)
+            }
+            other => other,
+        };
+
+        self.buf.push(func_def);
+
+        py_expr!("\n{func:id}", func = func_name.as_str())
+    }
+
+    fn lower_generator(&mut self, generator: ast::ExprGenerator) -> Expr {
+        let ast::ExprGenerator {
+            elt, generators, ..
+        } = generator;
+
+        let first_iter_expr = generators
+            .first()
+            .expect("generator expects at least one comprehension")
+            .iter
+            .clone();
+
+        let func_name = self.ctx.fresh("gen");
+
+        let param_name = if let Expr::Name(ast::ExprName { id, .. }) = &first_iter_expr {
+            id.clone()
         } else {
-            lowered_functions.push(stmt.clone());
-            *stmt = single_stmt(lowered_functions);
-            true
+            Name::new(self.ctx.fresh("iter"))
+        };
+
+        let mut body = vec![py_stmt!("\nyield {value:expr}", value = *elt)];
+
+        for comp in generators.iter().rev() {
+            let mut inner = body;
+            for if_expr in comp.ifs.iter().rev() {
+                inner = vec![py_stmt!(
+                    "\nif {test:expr}:\n    {body:stmt}",
+                    test = if_expr.clone(),
+                    body = inner,
+                )];
+            }
+            body = vec![if comp.is_async {
+                py_stmt!(
+                    "\nasync for {target:expr} in {iter:expr}:\n    {body:stmt}",
+                    target = comp.target.clone(),
+                    iter = comp.iter.clone(),
+                    body = inner,
+                )
+            } else {
+                py_stmt!(
+                    "\nfor {target:expr} in {iter:expr}:\n    {body:stmt}",
+                    target = comp.target.clone(),
+                    iter = comp.iter.clone(),
+                    body = inner,
+                )
+            }];
         }
+
+        if let Stmt::For(ast::StmtFor { iter, .. }) = body.first_mut().unwrap() {
+            *iter = Box::new(py_expr!("\n{name:id}", name = param_name.as_str()));
+        }
+
+        let func_def = py_stmt!(
+            "\ndef {func:id}({param:id}):\n    {body:stmt}",
+            func = func_name.as_str(),
+            param = param_name.as_str(),
+            body = body,
+        );
+
+        self.buf.push(func_def);
+
+        py_expr!(
+            "\n{func:id}(__dp__.iter({iter:expr}))",
+            iter = first_iter_expr,
+            func = func_name.as_str(),
+        )
     }
 
     fn rewrite_target(&mut self, target: Expr, value: Expr, out: &mut Vec<Stmt>) {
@@ -273,6 +366,8 @@ impl<'a> Transformer for ExprRewriter<'a> {
     fn visit_expr(&mut self, expr: &mut Expr) {
         let original = expr.clone();
         *expr = match original {
+            Expr::Lambda(lambda) => self.lower_lambda(lambda),
+            Expr::Generator(generator) => self.lower_generator(generator),
             Expr::FString(f_string) => rewrite_string::rewrite_fstring(f_string),
             Expr::TString(t_string) => rewrite_string::rewrite_tstring(t_string),
             Expr::Slice(ast::ExprSlice {
@@ -458,11 +553,6 @@ impl<'a> Transformer for ExprRewriter<'a> {
 
     fn visit_stmt(&mut self, stmt: &mut Stmt) {
         rewrite_complex_expr::rewrite(stmt, self.ctx);
-
-        if self.lower_lambdas_generators(stmt) {
-            self.visit_stmt(stmt);
-            return;
-        }
 
         if matches!(stmt, Stmt::FunctionDef(_)) {
             if let Stmt::FunctionDef(ast::StmtFunctionDef {
@@ -651,11 +741,6 @@ impl<'a> Transformer for ExprRewriter<'a> {
                 }
                 _ => {}
             }
-        }
-
-        if self.lower_lambdas_generators(stmt) {
-            self.visit_stmt(stmt);
-            return;
         }
     }
 }
