+++ src/transform/rewrite_expr_to_stmt.rs
@@ -1,161 +1,6 @@
-use std::cell::RefCell;
-
-use super::context::Context;
-use crate::body_transform::{walk_expr, walk_stmt, Transformer};
 use crate::template::{make_binop, make_unaryop, single_stmt};
 use crate::{py_expr, py_stmt};
-use ruff_python_ast::name::Name;
 use ruff_python_ast::{self as ast, CmpOp, Expr, Stmt};
-use ruff_text_size::TextRange;
-
-pub(crate) struct LambdaGeneratorLowerer<'ctx> {
-    ctx: &'ctx Context,
-    functions: RefCell<Vec<Stmt>>,
-}
-
-impl<'ctx> LambdaGeneratorLowerer<'ctx> {
-    pub(crate) fn new(ctx: &'ctx Context) -> Self {
-        Self {
-            ctx,
-            functions: RefCell::new(Vec::new()),
-        }
-    }
-
-    pub(crate) fn into_statements(self) -> Vec<Stmt> {
-        self.functions.into_inner()
-    }
-
-    pub(crate) fn rewrite(&mut self, stmt: &mut Stmt) {
-        walk_stmt(self, stmt);
-    }
-
-    fn lower_lambda(&mut self, expr: &mut Expr, lambda: ast::ExprLambda) {
-        let func_name = self.ctx.fresh("lambda");
-
-        let ast::ExprLambda {
-            parameters, body, ..
-        } = lambda;
-
-        let parameters = parameters
-            .map(|params| *params)
-            .unwrap_or_else(|| ast::Parameters {
-                range: TextRange::default(),
-                node_index: ast::AtomicNodeIndex::default(),
-                posonlyargs: vec![],
-                args: vec![],
-                vararg: None,
-                kwonlyargs: vec![],
-                kwarg: None,
-            });
-
-        let func_def = py_stmt!(
-            "\ndef {func:id}():\n    return {body:expr}",
-            func = func_name.as_str(),
-            body = *body,
-        );
-
-        self.functions.borrow_mut().push(match func_def {
-            Stmt::FunctionDef(mut function_def) => {
-                function_def.parameters = Box::new(parameters);
-                Stmt::FunctionDef(function_def)
-            }
-            other => other,
-        });
-
-        *expr = py_expr!("\n{func:id}", func = func_name.as_str());
-    }
-
-    fn lower_generator(&mut self, expr: &mut Expr, generator: ast::ExprGenerator) {
-        let ast::ExprGenerator {
-            elt, generators, ..
-        } = generator;
-
-        let first_iter_expr = generators
-            .first()
-            .expect("generator expects at least one comprehension")
-            .iter
-            .clone();
-
-        let func_name = self.ctx.fresh("gen");
-
-        let param_name = if let Expr::Name(ast::ExprName { id, .. }) = &first_iter_expr {
-            id.clone()
-        } else {
-            Name::new(self.ctx.fresh("iter"))
-        };
-
-        let mut body = vec![py_stmt!("\nyield {value:expr}", value = *elt)];
-
-        for comp in generators.iter().rev() {
-            let mut inner = body;
-            for if_expr in comp.ifs.iter().rev() {
-                inner = vec![py_stmt!(
-                    "\nif {test:expr}:\n    {body:stmt}",
-                    test = if_expr.clone(),
-                    body = inner,
-                )];
-            }
-            body = vec![if comp.is_async {
-                py_stmt!(
-                    "\nasync for {target:expr} in {iter:expr}:\n    {body:stmt}",
-                    target = comp.target.clone(),
-                    iter = comp.iter.clone(),
-                    body = inner,
-                )
-            } else {
-                py_stmt!(
-                    "\nfor {target:expr} in {iter:expr}:\n    {body:stmt}",
-                    target = comp.target.clone(),
-                    iter = comp.iter.clone(),
-                    body = inner,
-                )
-            }];
-        }
-
-        if let Stmt::For(ast::StmtFor { iter, .. }) = body.first_mut().unwrap() {
-            *iter = Box::new(py_expr!("\n{name:id}", name = param_name.as_str()));
-        }
-
-        let func_def = py_stmt!(
-            "\ndef {func:id}({param:id}):\n    {body:stmt}",
-            func = func_name.as_str(),
-            param = param_name.as_str(),
-            body = body,
-        );
-
-        self.functions.borrow_mut().push(func_def);
-
-        *expr = py_expr!(
-            "\n{func:id}(__dp__.iter({iter:expr}))",
-            iter = first_iter_expr,
-            func = func_name.as_str(),
-        );
-    }
-}
-
-impl<'ctx> Transformer for LambdaGeneratorLowerer<'ctx> {
-    fn visit_expr(&mut self, expr: &mut Expr) {
-        match expr.clone() {
-            Expr::Lambda(lambda) => {
-                self.lower_lambda(expr, lambda);
-                return;
-            }
-            Expr::Generator(generator) => {
-                self.lower_generator(expr, generator);
-                return;
-            }
-            _ => {}
-        }
-
-        walk_expr(self, expr);
-    }
-
-    fn visit_stmt(&mut self, _: &mut Stmt) {
-        // Only visit expressions referenced directly by this statement; callers
-        // are responsible for rewriting any nested statements themselves.
-    }
-}
-
 pub(crate) fn expr_boolop_to_stmts(target: &str, bool_op: ast::ExprBoolOp) -> Vec<Stmt> {
     let ast::ExprBoolOp { op, values, .. } = bool_op;
